refactor(value-detail): extract ConditionMatrix to its own file
DomainAnalysisValueDetail.tsx was 552 lines, over the 400-line limit.
Extract ConditionMatrix and its helpers into components/domains/ConditionMatrix.tsx
to bring both files under the limit (339 and 214 lines respectively).
No behaviour change.
